+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-# Version 3.8.1 (xxxx-xx-xx)
+# Version 3.8.1 (2019-02-??)
 
 * [new] Support for the `web.server.welcomeFiles` with Undertow. Default value is `index.html`.
+* [new] Support for file logging configuration when using LogBack. Enable by setting `logging.file.enabled` to `true`.  
 * [chg] Filter priorities updated: resource filter has -2000 priority, and Jersey filter has -1000 priority.
 
 # Version 3.8.0 (2018-11-27)